requirements.txt
spacy==3.7.4
stanza==1.7.0
pyyaml==6.0.1
tqdm==4.66.2
pytest==8.1.1
Wikipedia-API==0.6.0
beautifulsoup4==4.12.3
requests==2.32.3